[ROMM-49] Fix importing large firkmware files

+++ Models/RomM/Platform/RomMPlatformFirmware.cs
@@ -24,7 +24,7 @@ namespace RomM.Models.RomM.Platform
         public string FilePath { get; set; }
 
         [JsonProperty("file_size_bytes")]
-        public int FileSizeBytes { get; set; }
+        public ulong FileSizeBytes { get; set; }
 
         [JsonProperty("full_path")]
         public string FullPath { get; set; }
